requirements.txt
pymongo==4.11.1
python-dotenv==1.0.1
python-telegram-bot==22.0
Flask==3.1.0
gunicorn==23.0.0